requirements.txt
groq==0.15.0
python-dotenv==1.0.1
ipython==8.31.0
llama-index==0.12.11
llama-index-llms-groq==0.3.1
llama-index-embeddings-huggingface==0.5.0
docx2txt==0.8
